HNSW filtering implemented
