+++ core/src/idx/trees/hnsw/flavor.rs
@@ -1,0 +1,178 @@
+use crate::err::Error;
+use crate::idx::planner::checker::ConditionChecker;
+use crate::idx::trees::dynamicset::{ArraySet, HashBrownSet};
+use crate::idx::trees::hnsw::index::VecDocs;
+use crate::idx::trees::hnsw::{ElementId, Hnsw};
+use crate::idx::trees::vector::SharedVector;
+use crate::sql::index::HnswParams;
+use reblessive::tree::Stk;
+
+pub(super) type ASet<const N: usize> = ArraySet<ElementId, N>;
+pub(super) type HSet = HashBrownSet<ElementId>;
+
+pub(super) enum HnswFlavor {
+	H5_9(Hnsw<ASet<9>, ASet<5>>),
+	H5_17(Hnsw<ASet<17>, ASet<5>>),
+	H5_25(Hnsw<ASet<25>, ASet<5>>),
+	H5set(Hnsw<HSet, ASet<5>>),
+	H9_17(Hnsw<ASet<17>, ASet<9>>),
+	H9_25(Hnsw<ASet<25>, ASet<9>>),
+	H9set(Hnsw<HSet, ASet<9>>),
+	H13_25(Hnsw<ASet<25>, ASet<13>>),
+	H13set(Hnsw<HSet, ASet<13>>),
+	H17set(Hnsw<HSet, ASet<17>>),
+	H21set(Hnsw<HSet, ASet<21>>),
+	H25set(Hnsw<HSet, ASet<25>>),
+	H29set(Hnsw<HSet, ASet<29>>),
+	Hset(Hnsw<HSet, HSet>),
+}
+
+impl HnswFlavor {
+	pub(super) fn new(p: &HnswParams) -> Self {
+		match p.m {
+			1..=4 => match p.m0 {
+				1..=8 => Self::H5_9(Hnsw::<ASet<9>, ASet<5>>::new(p)),
+				9..=16 => Self::H5_17(Hnsw::<ASet<17>, ASet<5>>::new(p)),
+				17..=24 => Self::H5_25(Hnsw::<ASet<25>, ASet<5>>::new(p)),
+				_ => Self::H5set(Hnsw::<HSet, ASet<5>>::new(p)),
+			},
+			5..=8 => match p.m0 {
+				1..=16 => Self::H9_17(Hnsw::<ASet<17>, ASet<9>>::new(p)),
+				17..=24 => Self::H9_25(Hnsw::<ASet<25>, ASet<9>>::new(p)),
+				_ => Self::H9set(Hnsw::<HSet, ASet<9>>::new(p)),
+			},
+			9..=12 => match p.m0 {
+				17..=24 => Self::H13_25(Hnsw::<ASet<25>, ASet<13>>::new(p)),
+				_ => Self::H13set(Hnsw::<HSet, ASet<13>>::new(p)),
+			},
+			13..=16 => Self::H17set(Hnsw::<HSet, ASet<17>>::new(p)),
+			17..=20 => Self::H21set(Hnsw::<HSet, ASet<21>>::new(p)),
+			21..=24 => Self::H25set(Hnsw::<HSet, ASet<25>>::new(p)),
+			25..=28 => Self::H29set(Hnsw::<HSet, ASet<29>>::new(p)),
+			_ => Self::Hset(Hnsw::<HSet, HSet>::new(p)),
+		}
+	}
+
+	pub(super) fn insert(&mut self, q_pt: SharedVector) -> ElementId {
+		match self {
+			HnswFlavor::H5_9(h) => h.insert(q_pt),
+			HnswFlavor::H5_17(h) => h.insert(q_pt),
+			HnswFlavor::H5_25(h) => h.insert(q_pt),
+			HnswFlavor::H5set(h) => h.insert(q_pt),
+			HnswFlavor::H9_17(h) => h.insert(q_pt),
+			HnswFlavor::H9_25(h) => h.insert(q_pt),
+			HnswFlavor::H9set(h) => h.insert(q_pt),
+			HnswFlavor::H13_25(h) => h.insert(q_pt),
+			HnswFlavor::H13set(h) => h.insert(q_pt),
+			HnswFlavor::H17set(h) => h.insert(q_pt),
+			HnswFlavor::H21set(h) => h.insert(q_pt),
+			HnswFlavor::H25set(h) => h.insert(q_pt),
+			HnswFlavor::H29set(h) => h.insert(q_pt),
+			HnswFlavor::Hset(h) => h.insert(q_pt),
+		}
+	}
+	pub(super) fn remove(&mut self, e_id: ElementId) -> bool {
+		match self {
+			HnswFlavor::H5_9(h) => h.remove(e_id),
+			HnswFlavor::H5_17(h) => h.remove(e_id),
+			HnswFlavor::H5_25(h) => h.remove(e_id),
+			HnswFlavor::H5set(h) => h.remove(e_id),
+			HnswFlavor::H9_17(h) => h.remove(e_id),
+			HnswFlavor::H9_25(h) => h.remove(e_id),
+			HnswFlavor::H9set(h) => h.remove(e_id),
+			HnswFlavor::H13_25(h) => h.remove(e_id),
+			HnswFlavor::H13set(h) => h.remove(e_id),
+			HnswFlavor::H17set(h) => h.remove(e_id),
+			HnswFlavor::H21set(h) => h.remove(e_id),
+			HnswFlavor::H25set(h) => h.remove(e_id),
+			HnswFlavor::H29set(h) => h.remove(e_id),
+			HnswFlavor::Hset(h) => h.remove(e_id),
+		}
+	}
+	pub(super) fn knn_search(
+		&self,
+		pt: &SharedVector,
+		k: usize,
+		ef: usize,
+	) -> Vec<(f64, ElementId)> {
+		match self {
+			HnswFlavor::H5_9(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H5_17(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H5_25(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H5set(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H9_17(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H9_25(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H9set(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H13_25(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H13set(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H17set(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H21set(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H25set(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::H29set(h) => h.knn_search(pt, k, ef),
+			HnswFlavor::Hset(h) => h.knn_search(pt, k, ef),
+		}
+	}
+	pub(super) async fn knn_search_checked(
+		&self,
+		pt: &SharedVector,
+		k: usize,
+		ef: usize,
+		vec_docs: &VecDocs,
+		stk: &mut Stk,
+		chk: &mut ConditionChecker<'_>,
+	) -> Result<Vec<(f64, ElementId)>, Error> {
+		match self {
+			HnswFlavor::H5_9(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H5_17(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H5_25(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H5set(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H9_17(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H9_25(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H9set(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H13_25(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H13set(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H17set(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H21set(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H25set(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::H29set(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+			HnswFlavor::Hset(h) => h.knn_search_checked(pt, k, ef, vec_docs, stk, chk).await,
+		}
+	}
+	pub(super) fn get_vector(&self, e_id: &ElementId) -> Option<&SharedVector> {
+		match self {
+			HnswFlavor::H5_9(h) => h.get_vector(e_id),
+			HnswFlavor::H5_17(h) => h.get_vector(e_id),
+			HnswFlavor::H5_25(h) => h.get_vector(e_id),
+			HnswFlavor::H5set(h) => h.get_vector(e_id),
+			HnswFlavor::H9_17(h) => h.get_vector(e_id),
+			HnswFlavor::H9_25(h) => h.get_vector(e_id),
+			HnswFlavor::H9set(h) => h.get_vector(e_id),
+			HnswFlavor::H13_25(h) => h.get_vector(e_id),
+			HnswFlavor::H13set(h) => h.get_vector(e_id),
+			HnswFlavor::H17set(h) => h.get_vector(e_id),
+			HnswFlavor::H21set(h) => h.get_vector(e_id),
+			HnswFlavor::H25set(h) => h.get_vector(e_id),
+			HnswFlavor::H29set(h) => h.get_vector(e_id),
+			HnswFlavor::Hset(h) => h.get_vector(e_id),
+		}
+	}
+	#[cfg(test)]
+	pub(super) fn check_hnsw_properties(&self, expected_count: usize) {
+		match self {
+			HnswFlavor::H5_9(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H5_17(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H5_25(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H5set(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H9_17(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H9_25(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H9set(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H13_25(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H13set(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H17set(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H21set(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H25set(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::H29set(h) => h.check_hnsw_properties(expected_count),
+			HnswFlavor::Hset(h) => h.check_hnsw_properties(expected_count),
+		}
+	}
+}
